src/requirements.txt
flask >= 3.1.0
flask-cors >= 5.0.1
wtforms_json >= .3.5
